requirements_minimal.txt
fb2reader
beautifulsoup4
razdel
numpy
pandas
scikit-learn
pyyaml
mlflow
transformers[torch]
datasets
evaluate
tqdm
sacrebleu
nltk
rouge_score
absl-py
bert_score
--extra-index-url https://download.pytorch.org/whl/cu128
torch==2.7.0+cu128
